% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Week2functions.R
\name{compare_words}
\alias{compare_words}
\title{Compare Words}
\usage{
compare_words(guess_str)
}
\arguments{
\item{guess_str}{character. Should be a word with the same length as the target word}

\item{target_str}{character. Should be a word with the same length as the guessed word}
}
\value{
character vector. Shows whether the letter is in the word and on the correct place, in the word but incorrect place, or not in the word
}
\description{
This function compares whether the letters in the target word are the same as in the guessed word
}
\examples{
# Compare the guessed word with the target word
compare_words("hallo")
}
