% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/German Wordle.R
\name{playwordle}
\alias{playwordle}
\title{German Wordle}
\usage{
playwordle()
}
\arguments{
\item{/}{No input needed.}
}
\value{
A Shiny app interface for playing the German Word Guessing Game.
}
\description{
This is a function for playing the popular online-game Wordle in German in a Shiny App.
}
\examples{
playwordle()
}
