% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Week2functions.R
\name{compare_words}
\alias{compare_words}
\title{Compare Words}
\usage{
compare_words(target_str, guess_str)
}
\arguments{
\item{target_str}{character. Should be a word with the same length as the guessed word}

\item{guess_str}{character. Should be a word with the same length as the target word}
}
\value{
Shows whether the letter is in the word and on the correct place, in the word but incorrect place, or not in the word
}
\description{
This function compares whether the letters in the target word are the same as in the guessed word
}
\examples{
# compare_words("rauch","hallo")

# Compare two words
compare_words("rauch","hallo")
}
